% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDAG-package.R
\docType{package}
\name{CohortDAG-package}
\alias{CohortDAG}
\alias{CohortDAG-package}
\title{CohortDAG: DAG-Based Batch Cohort SQL Generation for OMOP CDM}
\description{
Generates optimized batch SQL for ATLAS/OHDSI cohort definitions using a directed acyclic graph (DAG) execution plan. Deduplicates shared computations across cohorts, translates ATLAS JSON to SQL without Java/CirceR, and supports incremental caching with Merkle-tree validation. Includes a library of 3000+ cohort definitions for benchmarking and validation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/darwin-eu/CohortDAG}
  \item Report bugs at \url{https://github.com/darwin-eu/CohortDAG/issues}
}

}
\author{
\strong{Maintainer}: Adam Black \email{ablack3@gmail.com} (\href{https://orcid.org/0000-0001-5576-8701}{ORCID})

Authors:
\itemize{
  \item Artem Gorbachev \email{artem.gorbachev@odysseusinc.com}
}

}
\keyword{internal}
